Seemed to have fixed the collision bug by moving the origin set to LoadContent() instead of individual constructor. Still need to test more just in case.

+++ E-Shooter/E-Shooter/E-Shooter/Abstracts/BulletObjectAbstract.cs
@@ -20,7 +20,7 @@ namespace E_Shooter
     {
 
         public static int fireCooldown;
-        public static int speed;
+        public static float speed;
         public static float rateCounter;
 
         public int damagePoints;
